added migrations and models to manage Commercial Module

+++ database/migrations/2018_02_22_150948_create_produits_table.php
@@ -17,14 +17,14 @@ class CreateProduitsTable extends Migration
             $table->increments('id');
             $table->string('nom')->nullable();
             $table->string('codeBarre')->nullable();
-            $table->float('prixAchat')->nullable();
-            $table->float('prixVente')->nullable();
+            $table->double('prixAchat')->nullable();
+            $table->double('prixVente')->nullable();
             $table->float('TVA')->nullable();
             $table->integer('addedBy');
             $table->boolean('estAchete')->nullable();
             $table->boolean('estVendu')->nullable();
             $table->string('description')->nullable();
-            $table->integer('categorie_id');
+            $table->unsignedInteger('categorie_id');
             $table->timestamps();
         });
     }
